requirements.txt
# alabaster==0.7.12
# argh==0.26.2
# argon2-cffi==20.1.0
# asn1crypto==1.4.0
# astroid==2.4.2
# astropy==4.0.2
# async-generator==1.10
# atomicwrites==1.4.0
# attrs==20.3.0
# autopep8==1.5.4
# Babel==2.8.1
# backcall==0.2.0
# backports.shutil-get-terminal-size==1.0.0
# bcrypt==3.2.0
# beautifulsoup4==4.9.3
# bitarray==1.6.1
# bkcharts==0.2
# bleach==3.2.1
# bokeh==2.2.3
# boto==2.49.0
# Bottleneck==1.3.2
# brotlipy==0.7.0
# certifi==2020.6.20
# cffi==1.14.3
# chardet==3.0.4
# click==7.1.2
# cloudpickle==1.6.0
# colorama==0.4.3
# comtypes==1.1.7
# contextlib2==0.6.0.post1
# cryptography==3.1.1
# cycler==0.10.0
# Cython==0.29.21
# cytoolz==0.11.0
# dask==2.30.0
# decorator==4.4.2
# defusedxml==0.6.0
# diff-match-patch==20200713
# distributed==2.30.1
# docutils==0.16
# entrypoints==0.3
# et-xmlfile==1.0.1
# fastcache==1.1.0
# filelock==3.0.12
# flake8==3.8.4
# Flask==1.1.2
# fsspec==0.8.3
# future==0.18.2
# gevent==20.9.0
# glob2==0.7
# greenlet==0.4.17
# h5py==2.10.0
# HeapDict==1.0.1
# html5lib==1.1
# idna==2.10
# imageio==2.9.0
# imagesize==1.2.0
# importlib-metadata==2.0.0
# iniconfig==1.1.1
# intervaltree==3.1.0
# ipykernel==5.3.4
# ipython==7.19.0
# ipython-genutils==0.2.0
# ipywidgets==7.5.1
# isort==4.3.21
# itsdangerous==1.1.0
# jdcal==1.4.1
# jedi==0.17.1
# Jinja2==2.11.2
# joblib==0.17.0
# json5==0.9.5
# jsonschema==3.2.0
# jupyter==1.0.0
# jupyter-client==6.1.7
# jupyter-console==6.2.0
# jupyter-core==4.6.3
# jupyterlab==2.2.6
# jupyterlab-pygments==0.1.2
# jupyterlab-server==1.2.0
# keyring==21.4.0
# kiwisolver==1.3.0
# lazy-object-proxy==1.4.3
# libarchive-c==2.9
# llvmlite==0.34.0
# locket==0.2.0
# lxml==4.6.1
# MarkupSafe==1.1.1
matplotlib==3.3.2
# mccabe==0.6.1
# mistune==0.8.4
# mock==4.0.2
# more-itertools==8.6.0
# mpmath==1.1.0
# msgpack==1.0.0
# multipledispatch==0.6.0
# nbclient==0.5.1
# nbconvert==6.0.7
# nbformat==5.0.8
# nest-asyncio==1.4.2
# networkx==2.5
# nltk==3.5
# nose==1.3.7
# notebook==6.1.4
# numba==0.51.2
# numexpr==2.7.1
numpy==1.17.0
# numpydoc==1.1.0
# olefile==0.46
# openpyxl==3.0.5
# packaging==20.4
pandas==1.1.3
# pandocfilters==1.4.3
# paramiko==2.7.2
# parso==0.7.0
# partd==1.1.0
path==15.0.0
pathlib2==2.3.5
pathtools==0.1.2
# patsy==0.5.1
# pep8==1.7.1
# pexpect==4.8.0
# pickleshare==0.7.5
# Pillow==8.0.1
# pip==20.2.4
# pkginfo==1.6.1
plotly==4.14.3
# pluggy==0.13.1
# ply==3.11
# prometheus-client==0.8.0
# prompt-toolkit==3.0.8
# psutil==5.7.2
# py==1.9.0
# pycodestyle==2.6.0
# pycosat==0.6.3
# pycparser==2.20
# pycrypto==2.6.1
# pydocstyle==5.1.1
# pyflakes==2.2.0
# Pygments==2.7.2
# pylint==2.5.3
# PyNaCl==1.4.0
# pyOpenSSL==19.1.0
# pyparsing==2.4.7
# pyreadline==2.1
# pyrsistent==0.17.3
# PySocks==1.7.1
# python-dateutil==2.8.1
# python-jsonrpc-server==0.4.0
# python-language-server==0.35.1
# pytz==2020.1
# PyWavelets==1.1.1
PyYAML==5.3.1
# pyzmq==19.0.2
# QDarkStyle==2.8.1
# QtAwesome==1.0.1
# qtconsole==4.7.7
# QtPy==1.9.0
# regex==2020.10.15
requests==2.24.0
# retrying==1.3.3
# rope==0.18.0
# ruamel-yaml==0.15.87
# scikit-image==0.17.2
# scikit-learn==0.23.2
# scipy==1.5.2
# seaborn==0.11.0
# Send2Trash==1.5.0
# simplegeneric==0.8.1
# singledispatch==3.4.0.3
# six==1.15.0
# snowballstemmer==2.0.0
# sortedcollections==1.2.1
# sortedcontainers==2.2.2
# soupsieve==2.0.1
# Sphinx==3.2.1
# sphinxcontrib-applehelp==1.0.2
# sphinxcontrib-devhelp==1.0.2
# sphinxcontrib-htmlhelp==1.0.3
# sphinxcontrib-jsmath==1.0.1
# sphinxcontrib-qthelp==1.0.3
# sphinxcontrib-serializinghtml==1.1.4
# sphinxcontrib-websupport==1.2.4
# spyder==4.1.5
# spyder-kernels==1.9.4
# SQLAlchemy==1.3.20
# statsmodels==0.12.0
# sympy==1.6.2
# tables==3.6.1
# tblib==1.7.0
# terminado==0.9.1
# testpath==0.4.4
# threadpoolctl==2.1.0
# tifffile==2020.10.1
# toml==0.10.1
# toolz==0.11.1
# tornado==6.0.4
# tqdm==4.50.2
# traitlets==5.0.5
# typed-ast==1.4.1
# typing-extensions==3.7.4.3
# ujson==4.0.1
# unicodecsv==0.14.1
# urllib3==1.25.11
# watchdog==0.10.3
# wcwidth==0.2.5
# webencodings==0.5.1
# Werkzeug==1.0.1
# wheel==0.35.1
# widgetsnbextension==3.5.1
# win-inet-pton==1.1.0
# win-unicode-console==0.5
# wincertstore==0.2
# wrapt==1.12.1
# xlrd==1.2.0
# XlsxWriter==1.3.7
# xlwt==1.3.0
# yapf==0.30.0
# zict==2.0.0
# zipp==3.4.0
# zope.event==4.5.0
# zope.interface==5.1.2
